+++ src/main/java/com/github/nthykier/debpkg/util/Deb822LocalQuickFixImpl.java
@@ -3,6 +3,7 @@ package com.github.nthykier.debpkg.util;
 import com.intellij.codeInspection.LocalQuickFix;
 import com.intellij.codeInspection.ProblemDescriptor;
 import com.intellij.codeInspection.util.IntentionFamilyName;
+import com.intellij.codeInspection.util.IntentionName;
 import com.intellij.openapi.project.Project;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -16,6 +17,11 @@ import java.util.function.BiConsumer;
 public class Deb822LocalQuickFixImpl implements LocalQuickFix {
 
 
+    @IntentionName
+    @NonNull
+    @Getter
+    private final String name;
+
     @IntentionFamilyName
     @NonNull
     @Getter
@@ -24,14 +30,13 @@ public class Deb822LocalQuickFixImpl implements LocalQuickFix {
     @NonNull
     private final BiConsumer<Project, ProblemDescriptor> fixCode;
 
-
     @Override
     public void applyFix(@NotNull Project project, @NotNull ProblemDescriptor descriptor) {
         fixCode.accept(project, descriptor);
     }
 
 
-    static Deb822LocalQuickFixImpl of(String baseName, BiConsumer<Project, ProblemDescriptor> fixCode) {
-        return new Deb822LocalQuickFixImpl(baseName, fixCode);
+    static Deb822LocalQuickFixImpl of(String name, BiConsumer<Project, ProblemDescriptor> fixCode) {
+        return new Deb822LocalQuickFixImpl(name, name, fixCode);
     }
 }