+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ## [Unreleased]
 ### Added
 - Add inspection for possibly incorrect architecture field value based on package name.
+- Add DM-Upload-Allowed as known field and an inspection suggesting removing it as it is obsolete.
 
 ### Changed
 - Internal/Chore: Update the CI dependency on `actions/setup-java`.